fix: statistics fix
